add migration state to nexus API
